+++ agate/table/compute.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
 
+from collections import OrderedDict
 from copy import copy
 
 from agate.rows import Row
@@ -7,7 +8,7 @@ from agate import utils
 
 
 @utils.allow_tableset_proxy
-def compute(self, computations):
+def compute(self, computations, replace=False):
     """
     Create a new table by applying one or more :class:`.Computation` instances
     to each row.
@@ -15,23 +16,45 @@ def compute(self, computations):
     :param computations:
         A sequence of pairs of new column names and :class:`.Computation`
         instances.
+    :param replace:
+        If :code:`True` then new column names can match existing names, and
+        those columns will be replaced with the computed data.
     :returns:
         A new :class:`.Table`.
     """
     column_names = list(copy(self.column_names))
     column_types = list(copy(self.column_types))
+    new_columns = OrderedDict()
 
     for new_column_name, computation in computations:
-        column_names.append(new_column_name)
-        column_types.append(computation.get_computed_data_type(self))
+        new_column_type = computation.get_computed_data_type(self)
+
+        if new_column_name in column_names:
+            if not replace:
+                raise ValueError('New column name "%s" already exists. Specify replace=True to replace with computed data.')
+
+            i = column_names.index(new_column_name)
+            column_types[i] = new_column_type
+        else:
+            column_names.append(new_column_name)
+            column_types.append(new_column_type)
 
         computation.validate(self)
 
-    new_columns = tuple(c.run(self) for n, c in computations)
+    for new_column_name, computation in computations:
+        new_columns[new_column_name] = computation.run(self)
+
     new_rows = []
 
     for i, row in enumerate(self.rows):
-        values = tuple(row) + tuple(c[i] for c in new_columns)
+        values = []
+
+        for j, column_name in enumerate(column_names):
+            if column_name in new_columns:
+                values.append(new_columns[column_name][i])
+            else:
+                values.append(row[j])
+
         new_rows.append(Row(values, column_names))
 
     return self._fork(new_rows, column_names, column_types)