+++ CHANGELOG.rst
@@ -1,6 +1,7 @@
 1.4.0
 -----
 
+* :meth:`.Table.compute` now accepts a :code:`replace` argument that allows columns to be overwritten. (#597)
 * :meth:`.Table.from_fixed` now creates an agate table from a fixed-width file. (#358)
 * :mod:`.fixed` now implements a general-purpose fixed-width file reader. (#358)
 * :class:`TypeTester` now correctly parses negative currency values as Number. (#595)